+++ bindings/matrix-sdk-ffi/CHANGELOG.md
@@ -29,7 +29,7 @@ All notable changes to this project will be documented in this file.
   This is primarily for Element X to give a dedicated error message in case
   it connects a homeserver with only this method available.
   ([#5222](https://github.com/matrix-org/matrix-rust-sdk/pull/5222))
-- Add `Action::NotifyInApp` and `RoomNotificationMode::MentionsAndKeywordsOnlyTheRestInApp` behind
+- Add `Action::NotifyInApp` and `RoomNotificationMode::PushMentionsAndKeywordsOnly` behind
   a new feature `unstable-msc3768`.
   ([#5441](https://github.com/matrix-org/matrix-rust-sdk/pull/5441))